Struct xmz_server::co_no2_kombisensor::kombisensor::kombisensor::Kombisensor

pub struct Kombisensor { /* fields omitted */ }

Platine des CO-NO2-Kombisensor-Mod

Methods

impl Kombisensor

fn new() -> Self

Erzeugt eine neue Instanz

fn get_modbus_slave_id(&self) -> u8

Liefert die Modbus Adresse des Kombisensors

Jede Platine, eine Kombisensor mit mehreren Sensormeßzellen, hat eine Modbus Slave ID die so genannte Modbus Adresse.

Examples

use xmz_server::*;

let kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_modbus_slave_id(), 247);

fn get_kombisensor_type(&self) -> String

Liefert den Typ des Kombisensors als String

Examples

use xmz_server::*;

let kombisensor = Kombisensor::new();
assert_eq!("RA-GAS GmbH CO/ NO2 Kombisensor Modbus", format!("{}", kombisensor.get_kombisensor_type()));

fn get_sensor(&self, num: usize) -> Option<&Sensor>

Liefert eine Sensor Referenz in Option verpackt

Parameters

num - Nummer des gesuchten Sensors

Examples

use xmz_server::*;

fn get_sensor_mut(&mut self, num: usize) -> Option<&mut Sensor>

Liefert eine mutable Sensor Referenz in Option verpackt

Parameters

num - Nummer des gesuchten Sensors

Examples

use xmz_server::*;

fn get_sensors(&self) -> &Vec<Sensor>

Liefert die Sensoren als Referenze zu einem Vector

Examples

use xmz_server::*;

fn get_sensors_mut(&mut self) -> &mut Vec<Sensor>

Liefert die Sensoren in einem mutablen Vector Slice

Examples

use xmz_server::*;

fn get_error_count(&self) -> u64

Liefert den Fehlerzähler zurück

Examples

use xmz_server::*;

let kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_error_count(), 0);

fn sensors_mut(&mut self) -> &mut Vec<Sensor>

Mutabler Iterrator FIXME: Doku fehlt

fn inc_error_count(&mut self)

Erhöht den Fehlerzähler um Eins

Examples

use xmz_server::*;

let mut kombisensor = Kombisensor::new();
assert_eq!(kombisensor.get_error_count(), 0);
kombisensor.inc_error_count();
assert_eq!(kombisensor.get_error_count(), 1);

fn reset_error_count(&mut self)

Setzt den Fehlerzähler auf Null zurück

Examples

use xmz_server::*;

let mut kombisensor = Kombisensor::new();
kombisensor.inc_error_count();
assert_eq!(kombisensor.get_error_count(), 1);
kombisensor.reset_error_count();
assert_eq!(kombisensor.get_error_count(), 0);

fn set_modbus_slave_id(&mut self, modbus_slave_id: u8)

Setzt die Modbus Slave Id

Examples

use xmz_server::*;

fn set_version(&mut self, version: String)

Setzt die Modbus Slave Id

Examples

use xmz_server::*;

fn set_sensors_and_values(&mut self,
sensor1_number: u16,
sensor2_number: u16,
sensor1_adc_value: u16,
sensor2_adc_value: u16,
sensor1_min_value: u16,
sensor2_min_value: u16,
sensor1_max_value: u16,
sensor2_max_value: u16,
sensor1_adc_value_at_nullgas: u16,
sensor2_adc_value_at_nullgas: u16,
sensor1_adc_value_at_messgas: u16,
sensor2_adc_value_at_messgas: u16,
sensor1_concentration_at_nullgas: u16,
sensor2_concentration_at_nullgas: u16,
sensor1_concentration_at_messgas: u16,
sensor2_concentration_at_messgas: u16,
sensor1_config: u16,
sensor2_config: u16)

Diese Funktion stellt erstmal sicher das jeder Kombisensor auch 2 Sensor Datenstructuren besitzt. Beim aller ersten Start des Systems z.B. besitzt jede Kombisensor Datenstruktur nur ein leeres Vec. Da die Sensor Werte meistens nicht in der Konfiguration zu finden sind.

Kein Witz, die Parameter dieser Funktion sind wirklich so zahlreich.

fn parse<'a>(&mut self, input: &'a [u8]) -> IResult<&'a [u8], ()>