Struct xmz_server::co_no2_kombisensor::sensor::sensor::Sensor

pub struct Sensor { /* fields omitted */ }

Representation der Firmware Daten einer Messzelle

Methods

impl Sensor

fn new() -> Self

Erzeugt eine neue Sensor Instanz

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn new_with_type(sensor_type: SensorType) -> Self

Erzeugt eine neue Sensor Instanz, eines bestimmten Sensor Typs

Attributes

sensor_type - SensorType Type der Messzelle

Examples

use xmz_server::*;

let sensor = Sensor::new_with_type(SensorType::SimulationNO2);

fn get_number(&self) -> u16

Liefert die Sensor Number

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_adc_value(&self) -> u16

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_min_value(&self) -> u16

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_max_value(&self) -> u16

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_adc_value_at_nullgas(&self) -> u16

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_adc_value_at_messgas(&self) -> u16

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_concentration_at_nullgas(&self) -> u16

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_concentration_at_messgas(&self) -> u16

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_sensor_type(&self) -> String

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_si(&self) -> String

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_config(&self) -> u16

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_mv(&self) -> u16

Liefert den berechneten milli Volt Wert

Examples

use xmz_server::*;

let sensor = Sensor::new();

fn get_concentration(&self) -> f64

Berechnet die Gaskonzentration mit einer linearen Funktion

Examples

use xmz_server::*;

let mut sensor = Sensor::new_with_type(SensorType::SimulationNO2);
sensor.set_adc_value_at_nullgas(114);
sensor.set_adc_value_at_messgas(875);
sensor.set_concentration_at_nullgas(0);
sensor.set_concentration_at_messgas(280);
sensor.set_adc_value(333);

assert_eq!(sensor.get_concentration(), 80.57818659658344);

fn set_adc_from_concentration(&mut self, concentration: f64)

Berechnet den ADC Wert aus einer gegebenen Konzentration

Examples

use xmz_server::*;

let mut sensor = Sensor::new_with_type(SensorType::SimulationNO2);
sensor.set_adc_value_at_nullgas(114);
sensor.set_adc_value_at_messgas(875);
sensor.set_concentration_at_nullgas(0);
sensor.set_concentration_at_messgas(280);
sensor.set_adc_from_concentration(80.57818659658344);

assert_eq!(sensor.get_adc_value(), 333);

fn set_adc_value(&mut self, value: u16)

Setzt den ADC Wert des Sensors

Examples

use xmz_server::*;

let mut sensor = Sensor::new();
assert_eq!(sensor.get_adc_value(), 0);
sensor.set_adc_value(100);
assert_eq!(sensor.get_adc_value(), 100);

fn set_number(&mut self, number: u16)

Setzt die Sensor Nummer

Examples

use xmz_server::*;

let mut sensor = Sensor::new();
assert_eq!(sensor.get_number(), 0);
sensor.set_number(100);
assert_eq!(sensor.get_number(), 100);

fn set_min_value(&mut self, min_value: u16)

Setzt minimalen Sensormesswert des Sensors

Examples

use xmz_server::*;

let mut sensor = Sensor::new();
assert_eq!(sensor.get_min_value(), 0);
sensor.set_min_value(10);
assert_eq!(sensor.get_min_value(), 10);

fn set_max_value(&mut self, max_value: u16)

Setzt den maximalen Sensormesswert des Sensors

Examples

use xmz_server::*;

let mut sensor = Sensor::new();
assert_eq!(sensor.get_max_value(), 0);
sensor.set_max_value(10);
assert_eq!(sensor.get_max_value(), 10);

fn set_adc_value_at_nullgas(&mut self, adc_value_at_nullgas: u16)

Setzt den Nullgas ADC Wert des Sensors

Examples

use xmz_server::*;

let mut sensor = Sensor::new();
assert_eq!(sensor.get_adc_value_at_nullgas(), 0);
sensor.set_adc_value_at_nullgas(100);
assert_eq!(sensor.get_adc_value_at_nullgas(), 100);

fn set_adc_value_at_messgas(&mut self, adc_value_at_messgas: u16)

Setzt den Messgas ADC Wert des Sensors

Examples

use xmz_server::*;

let mut sensor = Sensor::new();
assert_eq!(sensor.get_adc_value_at_messgas(), 0);
sensor.set_adc_value_at_messgas(100);
assert_eq!(sensor.get_adc_value_at_messgas(), 100);

fn set_concentration_at_nullgas(&mut self, concentration_at_nullgas: u16)

Setzt die Gaskonzentration des Nullgases

Examples

use xmz_server::*;

let mut sensor = Sensor::new();
assert_eq!(sensor.get_concentration_at_nullgas(), 0);
sensor.set_concentration_at_nullgas(100);
assert_eq!(sensor.get_concentration_at_nullgas(), 100);

fn set_concentration_at_messgas(&mut self, concentration_at_messgas: u16)

Setzt die Gaskonzentration des Messgases

Examples

use xmz_server::*;

let mut sensor = Sensor::new();
assert_eq!(sensor.get_concentration_at_messgas(), 0);
sensor.set_concentration_at_messgas(100);
assert_eq!(sensor.get_concentration_at_messgas(), 100);

fn set_config(&mut self, config: u16)

Setzt Configuration des Sensors

Examples

use xmz_server::*;

let mut sensor = Sensor::new();
assert_eq!(sensor.get_config(), 0);
sensor.set_config(1);
assert_eq!(sensor.get_config(), 1);

fn is_enabled(&self) -> bool

Liefert true oder false je nachdem ob der Sensor aktiv Ist

Examples

use xmz_server::*;

let sensor = Sensor::new_with_type(SensorType::SimulationNO2);
assert_eq!(sensor.is_enabled(), false);

fn update_adc_values_time(&mut self)

Update die Values, Timestamp Tuppel

fn get_adc_values_time(&self) -> &Vec<(u16, Instant)>

fn average(&mut self, duration: Duration) -> u16

fn weighten_average(&mut self, duration: Duration) -> u16

fn get_concentration_from_adc(&self, adc_value: &u16) -> f64