Function xmz_server::configuration::configuration::read_config_file

pub fn read_config_file() -> Result<String>

Diese Funktion liest die Konfigurationsdatei ein, je nach Umgebung

Entweder wird das programm im development Modus aufgerufen, hier wird die Konfigurationsdatei lokal gesucht und gelesen. Oder aber das Programm wird im produktiv Modus (not(feature = "development")) ausgeführt, in diesem wird zunächste /boot gemounted, anschließend die Konfigurationsdatei eingelesen und zum Schluss /boot umounted.