Module xmz_server::messzelle

Trait das eine einzelne Messzelle beschreibt

Eine einzelne Messzelle eines Sensors. Diese sitzt in der Regel auf einer Sensor Platine (sensor). Jeder Sensor hat mindestens eine Messzelle mit einem Wert und einem Mittelwert.

Re-exports

pub use self::metz_connect_analog_420::MetzConnectCI4Analog420;

pub use self::ra_gas_co_mod::RaGasCOMod;

pub use self::ra_gas_no2_mod::RaGasNO2Mod;

Modules

metz_connect_analog_420

ra_gas_co_mod

ra_gas_no2_mod

Enums

MesszelleError
Mögliche Fehler einer Messzelle

MesszelleType
Verfügbare Messzellen Typen

Constants

MAX_VALUES_FOR_N_MINUTES

Traits

Messzelle
Basis Trait das die Eigenschaften einer Messzelle beschreibt

Type Definitions

BoxedMesszelle

MesszelleList