Trait xmz_server::prelude::Error

pub trait Error: Debug + Display {
    fn description(&self) -> &str { ... }

    fn cause(&self) -> Option<&Error> { ... }
}

Error is a trait representing the basic expectations for error values, i.e. values of type E in Result<T, E>. Errors must describe themselves through the Display and [Debug] traits, and may provide cause chain information:

The cause method is generally used when errors cross "abstraction boundaries", i.e. when a one module must report an error that is "caused" by an error from a lower-level module. This setup makes it possible for the high-level module to provide its own errors that do not commit to any particular implementation, but also reveal some of its implementation for debugging via cause chains.

Provided Methods

fn description(&self) -> &str

This method is soft-deprecated.

Although using it won’t cause compilation warning, new code should use Display instead and new impls can omit it.

To obtain error description as a string, use to_string().

Examples

match "xc".parse::<u32>() {
    Err(e) => {
        // Print `e` itself, not `e.description()`.
        println!("Error: {}", e);
    }
    _ => println!("No error"),
}

fn cause(&self) -> Option<&Error>

The lower-level cause of this error, if any.

Examples

use std::error::Error;
use std::fmt;

#[derive(Debug)]
struct SuperError {
    side: SuperErrorSideKick,
}

impl fmt::Display for SuperError {
    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
        write!(f, "SuperError is here!")
    }
}

impl Error for SuperError {
    fn description(&self) -> &str {
        "I'm the superhero of errors"
    }

    fn cause(&self) -> Option<&Error> {
        Some(&self.side)
    }
}

#[derive(Debug)]
struct SuperErrorSideKick;

impl fmt::Display for SuperErrorSideKick {
    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
        write!(f, "SuperErrorSideKick is here!")
    }
}

impl Error for SuperErrorSideKick {
    fn description(&self) -> &str {
        "I'm SuperError side kick"
    }
}

fn get_super_error() -> Result<(), SuperError> {
    Err(SuperError { side: SuperErrorSideKick })
}

fn main() {
    match get_super_error() {
        Err(e) => {
            println!("Error: {}", e.description());
            println!("Caused by: {}", e.cause().unwrap());
        }
        _ => println!("No error"),
    }
}

Methods

impl Error + 'static

pub fn is<T>(&self) -> bool where
T: 'static + Error,

Returns true if the boxed type is the same as T

pub fn downcast_ref<T>(&self) -> Option<&T> where
T: 'static + Error,

Returns some reference to the boxed value if it is of type T, or None if it isn't.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T> where
T: 'static + Error,

Returns some mutable reference to the boxed value if it is of type T, or None if it isn't.

impl Error + 'static + Send

pub fn is<T>(&self) -> bool where
T: 'static + Error,

Forwards to the method defined on the type Any.

pub fn downcast_ref<T>(&self) -> Option<&T> where
T: 'static + Error,

Forwards to the method defined on the type Any.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T> where
T: 'static + Error,

Forwards to the method defined on the type Any.

impl Error + 'static + Sync + Send

pub fn is<T>(&self) -> bool where
T: 'static + Error,

Forwards to the method defined on the type Any.

pub fn downcast_ref<T>(&self) -> Option<&T> where
T: 'static + Error,

Forwards to the method defined on the type Any.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T> where
T: 'static + Error,

Forwards to the method defined on the type Any.

impl Error + 'static

pub fn downcast<T>(
self: Box<Error + 'static>
) -> Result<Box<T>, Box<Error + 'static>> where
T: 'static + Error,

Attempt to downcast the box to a concrete type.

impl Error + 'static + Send

pub fn downcast<T>(
self: Box<Error + 'static + Send>
) -> Result<Box<T>, Box<Error + 'static + Send>> where
T: 'static + Error,

Attempt to downcast the box to a concrete type.

impl Error + 'static + Sync + Send

pub fn downcast<T>(
self: Box<Error + 'static + Sync + Send>
) -> Result<Box<T>, Box<Error + 'static + Sync + Send>> where
T: 'static + Error,

Attempt to downcast the box to a concrete type.