Struct xmz_server::shift_register::shift_register::ShiftRegister

pub struct ShiftRegister {
    pub oe_pin: Option<u64>,
    pub ds_pin: Option<u64>,
    pub clock_pin: Option<u64>,
    pub latch_pin: Option<u64>,
    pub data: u64,
    // some fields omitted
}

Fields
oe_pin: Option<u64> ds_pin: Option<u64> clock_pin: Option<u64> latch_pin: Option<u64> data: u64
Methods

impl ShiftRegister

fn new(register_type: ShiftRegisterType) -> Self

Erzeugt ein neuen Shift Register

Arguments

register_type - Art des Shift Registers

Examples

use xmz_server::*;

let sim = ShiftRegister::new(ShiftRegisterType::Simulation);
assert_eq!(sim.data, 0b0);

fn set(&mut self, num: u64) -> Result<()>

Setzt das übergebene Bit im Shift Register data Buffer

Arguments

num - Nummer des zu setzenden Bits Diese Nummer ist Eins basiert!

Der Parameter ist nicht Null basiert. Das bedeutet set(1) setzt das erste Bit(0) im data Buffer.

Examples

use xmz_server::*;

let mut sim = ShiftRegister::new(ShiftRegisterType::Simulation);
assert_eq!(sim.data, 0b0);
sim.set(3);
assert_eq!(sim.data, 0b100);

More info: http://stackoverflow.com/questions/47981/how-do-you-set-clear-and-toggle-a-single-bit-in-c-c

fn get(&self, num: u64) -> bool

Abfrage ob ein Bit gesetzt ist, true wenn ja, false wenn das bit nicht gesetzt ist

Arguments

num - Nummer des abzufragenden Bits Diese Nummer ist Eins basiert!

Der Parameter ist nicht Null basiert. D.h. get(1) fragt das erste Bit(0) im data Buffer ab.

Examples

use xmz_server::*;

let mut sim = ShiftRegister::new(ShiftRegisterType::Simulation);
sim.set(1);
sim.set(3);
assert_eq!(sim.get(1), true);
assert_eq!(sim.get(2), false);
assert_eq!(sim.get(3), true);

More info: http://stackoverflow.com/questions/47981/how-do-you-set-clear-and-toggle-a-single-bit-in-c-c

fn clear(&mut self, num: u64) -> Result<()>

Löscht das übergebene Bit

Arguments

num - Nummer des zu löschenden Bits Diese Nummer ist Eins basiert!

Der Parameter ist nicht Null basiert. D.h. clear(1) löscht das erste Bit(0) im data Buffer.

Examples

use xmz_server::*;

let mut sim = ShiftRegister::new(ShiftRegisterType::Simulation);
assert_eq!(sim.data, 0b0);

sim.set(1);
sim.set(3);
assert_eq!(sim.get(1), true);
assert_eq!(sim.get(3), true);

sim.clear(3);
assert_eq!(sim.get(1), true);
assert_eq!(sim.get(3), false);

fn toggle(&mut self, num: u64) -> Result<()>

Schaltet das übergebene Bit um, war es Null dann wird es Eins und umgekehrt

Arguments

num - Nummer des zu wechselnden Bits Diese Nummer ist Eins basiert!

Der Parameter ist nicht Null basiert. D.h. toggle(1) schaltet das erste Bit(0) im data Buffer um.

Examples

use xmz_server::*;

let mut sim = ShiftRegister::new(ShiftRegisterType::Simulation);
assert_eq!(sim.data, 0b0);

sim.toggle(3);
assert_eq!(sim.get(3), true);
sim.toggle(3);
assert_eq!(sim.get(3), false);

fn reset(&mut self) -> Result<()>

Reset nullt den Datenspeicher und gleicht ihn mit der Hardware ab.

Examples

use xmz_server::*;

let mut sim = ShiftRegister::new(ShiftRegisterType::Simulation);
assert_eq!(sim.get(1), false);
sim.set(1);
assert_eq!(sim.get(1), true);
sim.reset();
assert_eq!(sim.get(1), false);

fn test(&mut self) -> Result<()>

Lampentest testet alle Outputs

Diese Funktion schaltet alle Ausgänge high, wartet eine Sekunde und schaltet danach alle Ausgänge wieder aus.

Examples

use xmz_server::*;

let mut sim = ShiftRegister::new(ShiftRegisterType::Simulation);

sim.set(1);
sim.clear(10);
sim.test();
assert_eq!(sim.get(1), true);
assert_eq!(sim.get(10), false);

fn test_random(&mut self) -> Result<()>

Random Lampentest testet einige, wirklich vorhanden, Outputs, zufällig

Diese Funktion schaltet alle Ausgänge high, wartet eine Sekunde und schaltet danach alle Ausgänge wieder aus.

Examples

use xmz_server::*;

let mut sim = ShiftRegister::new(ShiftRegisterType::Simulation);

sim.test_random();